Public_Experiment/requirements.txt
azure-cognitiveservices-speech == 1.17.0
keyboard == 0.13.5
xlrd == 2.0.1
openpyxl == 3.0.7
tensorflow == 2.6.0
keras == 2.6.0
pandas == 1.3.4